Added deleting and updating sets, fixes

+++ src/main/java/org/staniszewska/dresscode/controller/ElementController.java
@@ -7,6 +7,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import org.staniszewska.dresscode.model.ElementDTO;
+import org.staniszewska.dresscode.repository.ElementRepository;
 import org.staniszewska.dresscode.service.ElementService;
 
 import java.util.List;
@@ -17,29 +18,31 @@ import java.util.List;
 public class ElementController {
     private final Logger logger = LogManager.getLogger(ElementController.class);
     private final ElementService elementService;
+    private final ElementRepository elementRepository;
 
     @GetMapping
     public ResponseEntity<List<ElementDTO>> getElements() {
         var result = elementService.findAll();
         if (result.isEmpty()) {
             logger.info("No elements found!");
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         } else {
             logger.info("All elements:");
             result.forEach((e) -> logger.info(e.toString()));
+            return new ResponseEntity<>(elementService.findAll(), HttpStatus.OK);
         }
-        return new ResponseEntity<>(elementService.findAll(), HttpStatus.OK);
     }
 
     @GetMapping(path = "/{id}")
-    public ResponseEntity<ElementDTO> getElementById(@PathVariable("id") Long id) {
+    public ResponseEntity<ElementDTO> getElementById(@PathVariable Long id) {
         var result = elementService.findById(id);
         if (result == null) {
             logger.info("Element not found!!");
-            return new ResponseEntity<>(elementService.findById(id), HttpStatus.OK);
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         } else {
             logger.info("Found element:");
             logger.info(result.toString());
-            return null;
+            return new ResponseEntity<>(elementService.findById(id), HttpStatus.OK);
         }
     }
 
@@ -50,15 +53,20 @@ public class ElementController {
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
-    @DeleteMapping(path = "/{id}/delete", produces = "application/json")
-    public ResponseEntity<Void> deleteElement(@PathVariable("id") Long id){
+    @DeleteMapping(path = "/delete/{id}", produces = "application/json")
+    public ResponseEntity<Void> deleteElement(@PathVariable Long id){
         logger.info("Deleting element...");
-        elementService.deleteElement(id);
-        return new ResponseEntity<>(HttpStatus.OK);
+        var result = elementService.findById(id);
+        if (result == null){
+            logger.info("Element not found!!");
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }else{
+            elementService.deleteElement(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
     }
 
-
-    @PutMapping(path = "/{id}/update", produces = "application/json")
+    @PutMapping(path = "/update/{id}", produces = "application/json")
     public ResponseEntity<Void> updateElement(@PathVariable Long id, @RequestBody ElementDTO elementDTO){
         logger.info("Updating element with id {}...", id);
         elementService.updateElement(id, elementDTO);
